clean out source maps for build
