+++ lib/apply/applyDiff.js
@@ -67,4 +67,3 @@ function applyDiff(context) {
     return context;
   });
 }
-//# sourceMappingURL=applyDiff.js.map
