+++ lib/apply/index.js
@@ -68,4 +68,3 @@ function index() {
 
   return (0, _prepareAll2.default)(filename, options).then(ensure).then((0, _diff2.default)('rules')).then((0, _diff2.default)('connections')).then(_ramda2.default.ifElse(isDryRun, _printApply2.default, _applyDiff2.default)).catch(error);
 }
-//# sourceMappingURL=index.js.map
