+++ lib/apply/loadManifest.js
@@ -38,4 +38,3 @@ function loadManifest(filename) {
 
   return _ramda2.default.assoc('manifest', _ramda2.default.evolve(transformations, manifest));
 }
-//# sourceMappingURL=loadManifest.js.map
