+++ lib/apply/printApply.js
@@ -92,4 +92,3 @@ function printDiff(context) {
 
   return context;
 }
-//# sourceMappingURL=printApply.js.map
