+++ lib/auth0/clients/fetchClients.js
@@ -17,4 +17,3 @@ function fetchClients(context) {
 
   return client.clients.getAll().then(_ramda2.default.assoc('clients', _ramda2.default.__, context));
 }
-//# sourceMappingURL=fetchClients.js.map
