+++ lib/auth0/clients/removeClients.js
@@ -25,4 +25,3 @@ function removeClients(context) {
     return context;
   });
 }
-//# sourceMappingURL=removeClients.js.map
