+++ lib/auth0/clients/updateClients.js
@@ -37,4 +37,3 @@ function updateClients(context) {
     return context;
   });
 }
-//# sourceMappingURL=updateClients.js.map
