+++ lib/auth0/connections/addConnections.js
@@ -32,4 +32,3 @@ function addConnections(context) {
     return context;
   });
 }
-//# sourceMappingURL=addConnections.js.map
