+++ lib/auth0/connections/fetchConnections.js
@@ -17,4 +17,3 @@ function fetchRules(context) {
 
   return client.connections.getAll().then(_ramda2.default.assoc('connections', _ramda2.default.__, context));
 }
-//# sourceMappingURL=fetchConnections.js.map
