+++ lib/auth0/connections/removeConnections.js
@@ -25,4 +25,3 @@ function removeConnections(context) {
     return context;
   });
 }
-//# sourceMappingURL=removeConnections.js.map
