+++ lib/auth0/connections/updateConnections.js
@@ -37,4 +37,3 @@ function updateConnections(context) {
     return context;
   });
 }
-//# sourceMappingURL=updateConnections.js.map
