+++ lib/auth0/rules/addRules.js
@@ -32,4 +32,3 @@ function addRules(context) {
     return context;
   });
 }
-//# sourceMappingURL=addRules.js.map
