+++ lib/auth0/rules/fetchRules.js
@@ -21,4 +21,3 @@ function fetchRules(context) {
     return client.rules.getAll({ stage: stage });
   }, STAGES)).then(_ramda2.default.flatten).then(_ramda2.default.assoc('rules', _ramda2.default.__, context));
 }
-//# sourceMappingURL=fetchRules.js.map
