+++ lib/auth0/rules/removeRules.js
@@ -25,4 +25,3 @@ function removeRules(context) {
     return context;
   });
 }
-//# sourceMappingURL=removeRules.js.map
