+++ lib/auth0/rules/updateRules.js
@@ -37,4 +37,3 @@ function updateRules(context) {
     return context;
   });
 }
-//# sourceMappingURL=updateRules.js.map
