+++ lib/cli.js
@@ -83,4 +83,3 @@ exports.cli = function (args) {
     return cli.parse(args);
   }).catch(console.error);
 };
-//# sourceMappingURL=cli.js.map
