+++ lib/config/configManager.js
@@ -97,4 +97,3 @@ function configManager() {
     }
   };
 }
-//# sourceMappingURL=configManager.js.map
