+++ lib/config/index.js
@@ -68,4 +68,3 @@ function index(command, args, options) {
 
   return (0, _loadEnv2.default)(options).then((0, _ramda.partial)(action, [args])).catch(error);
 }
-//# sourceMappingURL=index.js.map
