+++ lib/diff/detectors/additions.js
@@ -12,4 +12,3 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 var additions = _ramda2.default.differenceWith(_ramda2.default.eqProps('name'));
 exports.default = additions;
-//# sourceMappingURL=additions.js.map
