+++ lib/diff/detectors/changes.js
@@ -54,4 +54,3 @@ var changes = function changes(filter) {
 };
 
 exports.default = changes;
-//# sourceMappingURL=changes.js.map
