+++ lib/diff/detectors/removals.js
@@ -16,4 +16,3 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 var removals = _ramda2.default.flip(_additions2.default);
 exports.default = removals;
-//# sourceMappingURL=removals.js.map
