+++ lib/diff/diff.js
@@ -55,4 +55,3 @@ function diff(field, context) {
 }
 
 exports.default = _ramda2.default.curry(diff);
-//# sourceMappingURL=diff.js.map
