+++ lib/diff/index.js
@@ -25,9 +25,8 @@ function index() {
   var filename = arguments.length <= 0 || arguments[0] === undefined ? './auth0.yml' : arguments[0];
   var options = arguments[1];
 
-  return (0, _prepareAll2.default)(filename, options).then((0, _diff2.default)('rules')).then(function (_ref) {
+  return (0, _prepareAll2.default)(filename, options).then((0, _diff2.default)('rules')).then((0, _diff2.default)('connections')).then(function (_ref) {
     var diff = _ref.diff;
     return _ramda2.default.compose(console.log, formatJSON)(diff);
   });
 }
-//# sourceMappingURL=index.js.map
