+++ lib/dump/index.js
@@ -29,7 +29,9 @@ var _fetchClients = require('../auth0/clients/fetchClients');
 
 var _fetchClients2 = _interopRequireDefault(_fetchClients);
 
-var _associateClientConnections = require('../utils/associateClientConnections');
+var _associate = require('../utils/clientConnections/associate');
+
+var _associate2 = _interopRequireDefault(_associate);
 
 var _saveRules = require('./saveRules');
 
@@ -57,6 +59,5 @@ function index(filePath, options) {
   var ext = format == "json" ? "json" : "yml";
   var filename = filePath ? filePath : 'auth0.' + ext;
 
-  return (0, _loadEnv2.default)(options).then(_setupClient2.default).then(_fetchRules2.default).then(_fetchConnections2.default).then(_fetchClients2.default).then(_saveRules2.default).then(_saveConnections2.default).then(_associateClientConnections.associate).then((0, _saveManifest2.default)(filename)).then(ok('Manifest written to ' + filename)).catch(error);
+  return (0, _loadEnv2.default)(options).then(_setupClient2.default).then(_fetchRules2.default).then(_fetchConnections2.default).then(_fetchClients2.default).then(_saveRules2.default).then(_saveConnections2.default).then(_associate2.default).then((0, _saveManifest2.default)(filename)).then(ok('Manifest written to ' + filename)).catch(error);
 }
-//# sourceMappingURL=index.js.map
