+++ lib/dump/saveConnections.js
@@ -43,4 +43,3 @@ function saveConnections(context) {
 
   return _ramda2.default.assoc('connections', saveAndWrite(connections), context);
 }
-//# sourceMappingURL=saveConnections.js.map
