+++ lib/dump/saveManifest.js
@@ -54,4 +54,3 @@ function saveManifest(filename) {
     return _ramda2.default.compose(writer, formatter(format), _ramda2.default.evolve(transformations), selectTypes)(context).then(context);
   };
 }
-//# sourceMappingURL=saveManifest.js.map
