+++ lib/dump/saveRules.js
@@ -23,4 +23,3 @@ function saveRules(context) {
 
   return _ramda2.default.assoc('rules', _ramda2.default.map(saveScriptTo, rules), context);
 }
-//# sourceMappingURL=saveRules.js.map
