+++ lib/jwt/generateJWT.js
@@ -51,4 +51,3 @@ function generateJWT(context) {
     jti: uid
   }, decodedSecret);
 }
-//# sourceMappingURL=generateJWT.js.map
