+++ lib/jwt/index.js
@@ -28,4 +28,3 @@ function index(options) {
 
   return (0, _loadEnv2.default)(options).then(_generateJWT2.default).then(ok).catch(error);
 }
-//# sourceMappingURL=index.js.map
