+++ lib/setup/index.js
@@ -49,4 +49,3 @@ function index(options) {
     return ok('New JWT generated: ' + jwt);
   }).catch(error);
 }
-//# sourceMappingURL=index.js.map
