+++ lib/transformers/clients/prepareClientForDiff.js
@@ -12,4 +12,3 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 var prepareClientForDiff = _ramda2.default.pick(['name']);
 exports.default = prepareClientForDiff;
-//# sourceMappingURL=prepareClientForDiff.js.map
