+++ lib/transformers/clients/prepareClientForSave.js
@@ -12,4 +12,3 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 var prepareClientForSave = _ramda2.default.pick(['name']);
 exports.default = prepareClientForSave;
-//# sourceMappingURL=prepareClientForSave.js.map
