+++ lib/transformers/clients/prepareClientForUpdate.js
@@ -15,4 +15,3 @@ var ignoreFields = _ramda2.default.omit(['client_id', 'callback_url_template', '
 var prepareClientForUpdate = _ramda2.default.compose(filterNulls, ignoreFields, _ramda2.default.head);
 
 exports.default = prepareClientForUpdate;
-//# sourceMappingURL=prepareClientForUpdate.js.map
