+++ lib/transformers/connections/loadConnections.js
@@ -21,4 +21,3 @@ var loadConnectionScripts = _ramda2.default.over(_ramda2.default.lensPath(['opti
 
 var loadConnections = _ramda2.default.map(loadConnectionScripts);
 exports.default = loadConnections;
-//# sourceMappingURL=loadConnections.js.map
