+++ lib/transformers/connections/prepareConnectionForDiff.js
@@ -15,4 +15,3 @@ var filterOptions = _ramda2.default.over(_ramda2.default.lensProp('options'), _r
 var prepareConnectionForDiff = _ramda2.default.compose(filterOptions, filterFields);
 
 exports.default = prepareConnectionForDiff;
-//# sourceMappingURL=prepareConnectionForDiff.js.map
