+++ lib/transformers/connections/prepareConnectionForSave.js
@@ -15,4 +15,3 @@ var filterOptions = _ramda2.default.over(_ramda2.default.lensProp('options'), _r
 var prepareConnectionForSave = _ramda2.default.compose(filterOptions, filterFields);
 
 exports.default = prepareConnectionForSave;
-//# sourceMappingURL=prepareConnectionForSave.js.map
