+++ lib/transformers/connections/prepareConnectionForUpdate.js
@@ -14,4 +14,3 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 var filterFields = _ramda2.default.omit(['id', 'name', 'strategy']);
 var prepareConnectionForUpdate = _ramda2.default.compose(filterFields, _ramda2.default.head);
 exports.default = prepareConnectionForUpdate;
-//# sourceMappingURL=prepareConnectionForUpdate.js.map
