+++ lib/transformers/objectManipulation.js
@@ -31,4 +31,3 @@ function combineObject(keyName, valueName) {
 
 exports.expandObject = expandObject;
 exports.combineObject = combineObject;
-//# sourceMappingURL=objectManipulation.js.map
