+++ lib/transformers/rules/loadRules.js
@@ -16,4 +16,3 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 var loadRules = _ramda2.default.map((0, _loadScript2.default)('script'));
 exports.default = loadRules;
-//# sourceMappingURL=loadRules.js.map
