+++ lib/transformers/rules/prepareRuleForDiff.js
@@ -12,4 +12,3 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 var prepareRuleForDiff = _ramda2.default.pick(['script', 'stage', 'enabled', 'name', 'order']);
 exports.default = prepareRuleForDiff;
-//# sourceMappingURL=prepareRuleForDiff.js.map
