+++ lib/transformers/rules/prepareRuleForUpdate.js
@@ -13,4 +13,3 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 var filterFields = _ramda2.default.omit(['id', 'stage']);
 var prepareRuleForUpdate = _ramda2.default.compose(filterFields, _ramda2.default.head);
 exports.default = prepareRuleForUpdate;
-//# sourceMappingURL=prepareRuleForUpdate.js.map
