+++ lib/utils/apiErrorHandler.js
@@ -13,4 +13,3 @@ function apiErrorHandler(obj, actionDesc, context) {
     process.exit(1);
   };
 }
-//# sourceMappingURL=apiErrorHandler.js.map
