+++ lib/utils/loadEnv.js
@@ -18,4 +18,3 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 function loadEnv(options) {
   return Promise.resolve({ config: (0, _configManager2.default)(), options: options, say: (0, _say2.default)(options) });
 }
-//# sourceMappingURL=loadEnv.js.map
