+++ lib/utils/loadScript.js
@@ -23,4 +23,3 @@ function loadScript(field, object) {
 }
 
 exports.default = _ramda2.default.curry(loadScript);
-//# sourceMappingURL=loadScript.js.map
