+++ lib/utils/prepareAll.js
@@ -29,11 +29,12 @@ var _fetchClients = require('../auth0/clients/fetchClients');
 
 var _fetchClients2 = _interopRequireDefault(_fetchClients);
 
-var _associateClientConnections = require('../utils/associateClientConnections');
+var _disassociate = require('./clientConnections/disassociate');
+
+var _disassociate2 = _interopRequireDefault(_disassociate);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function prepareAll(filename, options) {
-  return (0, _loadEnv2.default)(options).then((0, _loadManifest2.default)(filename)).then(_setupClient2.default).then(_fetchRules2.default).then(_fetchConnections2.default).then(_fetchClients2.default).then(_associateClientConnections.disassociate);
+  return (0, _loadEnv2.default)(options).then((0, _loadManifest2.default)(filename)).then(_setupClient2.default).then(_fetchRules2.default).then(_fetchConnections2.default).then(_fetchClients2.default).then(_disassociate2.default);
 }
-//# sourceMappingURL=prepareAll.js.map
