+++ lib/utils/promisify.js
@@ -23,4 +23,3 @@ function promisify(fn) {
     });
   };
 }
-//# sourceMappingURL=promisify.js.map
