+++ lib/utils/saveScripts.js
@@ -34,4 +34,3 @@ function saveScripts(dir) {
     return _ramda2.default.assoc('script', path, scriptDefn);
   };
 }
-//# sourceMappingURL=saveScripts.js.map
