+++ lib/utils/say.js
@@ -58,4 +58,3 @@ function say(options) {
     return _ramda2.default.compose(boundPrint, fmt, hydrate);
   }, printers);
 }
-//# sourceMappingURL=say.js.map
