+++ lib/utils/setupClient.js
@@ -34,4 +34,3 @@ var _generateJWT2 = _interopRequireDefault(_generateJWT);
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 var ManagementClient = _auth2.default.ManagementClient;
-//# sourceMappingURL=setupClient.js.map
