+++ lib/utils/versionParser.js
@@ -20,4 +20,3 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 function versionParser() {
   return (0, _promisify2.default)(_fs2.default.readFile)(__dirname + '/../../package.json').then(JSON.parse).then((0, _ramda.path)(['version']));
 }
-//# sourceMappingURL=versionParser.js.map
